cockpit.js: split base64 helpers to their own file
Another straight move (with a minor indentation change to pacify
eslint).
